Refactor generation code
