+++ tessellation/procgen/tessellation_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class TessellationType(str, Enum):
+    """Enum representing the possible tesselation types."""
+
+    SQUARE_TRANSLATION = "SQUARE_TRANSLATION"
